channel parties are now ordered
